+++ src/watch.tsx
@@ -1,7 +1,4 @@
-/**
- * @function Watch
- */
-import * as React from 'react';
+import React, { useEffect, useState } from 'react';
 
 import GithubButton from './lib/githubButton';
 
@@ -11,37 +8,29 @@ export interface IPropTypes {
   color?: string;
 }
 
-export interface IState {
-  watchers_count: number;
-}
+const Watch: React.FC<IPropTypes> = (props) => {
+  const { owner, repo, color = '#6a737d' } = props;
+  const [watchersCount, setWatchersCount] = useState(0);
 
-export default class Watch extends React.Component<IPropTypes, IState> {
-  constructor(props: IPropTypes) {
-    super(props);
-    this.state = {
-      watchers_count: 0,
+  useEffect(() => {
+    const getCount = () => {
+      fetch(`https://api.github.com/repos/${owner}/${repo}`)
+        .then((res) => res.json())
+        .then((res) => {
+          setWatchersCount(res.watchers_count);
+        });
     };
-  }
+    getCount();
+  }, [owner, repo]);
 
-  componentDidMount() {
-    const { owner, repo } = this.props;
-    fetch(`https://api.github.com/repos/${owner}/${repo}`)
-      .then((res) => res.json())
-      .then((res) => {
-        this.setState({ watchers_count: res.watchers_count });
-      });
-  }
+  return (
+    <GithubButton
+      variant="watch"
+      count={watchersCount}
+      color={color}
+      {...props}
+    />
+  );
+};
 
-  render() {
-    const { color = '#6a737d' } = this.props;
-    const { watchers_count } = this.state;
-    return (
-      <GithubButton
-        variant="watch"
-        count={watchers_count}
-        color={color}
-        {...this.props}
-      />
-    );
-  }
-}
+export default React.memo(Watch);